Move access token related code to a separate package
