+++ app/src/main/kotlin/com/waz/zclient/core/di/NetworkModule.kt
@@ -5,16 +5,16 @@ import com.waz.zclient.BuildConfig
 import com.waz.zclient.core.di.NetworkDependencyProvider.createHttpClient
 import com.waz.zclient.core.di.NetworkDependencyProvider.createHttpClientForToken
 import com.waz.zclient.core.di.NetworkDependencyProvider.retrofit
-import com.waz.zclient.core.network.AccessTokenAuthenticator
-import com.waz.zclient.core.network.AccessTokenInterceptor
-import com.waz.zclient.core.network.AccessTokenLocalDataSource
-import com.waz.zclient.core.network.AccessTokenMapper
-import com.waz.zclient.core.network.AccessTokenRemoteDataSource
-import com.waz.zclient.core.network.AccessTokenRepository
+import com.waz.zclient.core.network.accesstoken.AccessTokenAuthenticator
+import com.waz.zclient.core.network.accesstoken.AccessTokenInterceptor
+import com.waz.zclient.core.network.accesstoken.AccessTokenLocalDataSource
+import com.waz.zclient.core.network.accesstoken.AccessTokenMapper
+import com.waz.zclient.core.network.accesstoken.AccessTokenRemoteDataSource
+import com.waz.zclient.core.network.accesstoken.AccessTokenRepository
 import com.waz.zclient.core.network.ApiService
 import com.waz.zclient.core.network.NetworkClient
 import com.waz.zclient.core.network.NetworkHandler
-import com.waz.zclient.core.network.RefreshTokenMapper
+import com.waz.zclient.core.network.accesstoken.RefreshTokenMapper
 import com.waz.zclient.core.network.RetrofitClient
 import com.waz.zclient.core.network.api.token.TokenApi
 import com.waz.zclient.core.network.api.token.TokenService
