+++ app/src/main/kotlin/com/waz/zclient/core/network/accesstoken/AccessToken.kt
@@ -1,4 +1,4 @@
-package com.waz.zclient.core.network
+package com.waz.zclient.core.network.accesstoken
 
 import com.waz.zclient.core.extension.empty
 import com.waz.zclient.core.network.api.token.AccessTokenResponse
