+++ app/src/main/kotlin/com/waz/zclient/core/network/accesstoken/AccessTokenInterceptor.kt
@@ -1,4 +1,4 @@
-package com.waz.zclient.core.network
+package com.waz.zclient.core.network.accesstoken
 
 import okhttp3.Interceptor
 import okhttp3.Response
