+++ app/src/main/kotlin/com/waz/zclient/core/network/accesstoken/AccessTokenPreference.kt
@@ -1,4 +1,4 @@
-package com.waz.zclient.core.network
+package com.waz.zclient.core.network.accesstoken
 
 import com.google.gson.annotations.SerializedName
