+++ app/src/main/kotlin/com/waz/zclient/core/network/accesstoken/AccessTokenRepository.kt
@@ -1,4 +1,4 @@
-package com.waz.zclient.core.network
+package com.waz.zclient.core.network.accesstoken
 
 import com.waz.zclient.core.exception.Failure
 import com.waz.zclient.core.functional.Either
