+++ app/src/test/kotlin/com/waz/zclient/core/network/accesstoken/AccessTokenMapperTest.kt
@@ -1,4 +1,4 @@
-package com.waz.zclient.core.network
+package com.waz.zclient.core.network.accesstoken
 
 import com.waz.zclient.UnitTest
 import com.waz.zclient.core.network.api.token.AccessTokenResponse
