+++ app/src/test/kotlin/com/waz/zclient/core/network/accesstoken/AccessTokenRemoteDataSourceTest.kt
@@ -1,6 +1,7 @@
-package com.waz.zclient.core.network
+package com.waz.zclient.core.network.accesstoken
 
 import com.waz.zclient.UnitTest
+import com.waz.zclient.core.network.accesstoken.AccessTokenRemoteDataSource
 import com.waz.zclient.core.network.api.token.TokenService
 import org.junit.Before
 import org.junit.Test
